Refactor OpenAI provider into dedicated module
